Add files via upload
This assignment requires you to design and implement a Book class with attributes for title, author, ISBN, and price. It also requires a method to find the number of created Book objects, and to compare Book objects for equality based on ISBN and price. Additionally, you need to create a driver program with a main menu offering options to enter new books, change book information, display books by author, display books under a certain price, and quit. It includes password protection for certain actions and limits on failed attempts to prevent suspicious activities.